+++ src/OrigoDB.Core/Utilities/Extensions.cs
@@ -9,12 +9,12 @@ namespace OrigoDB.Core
     public static class Extensions
     {
 
-        public static Int64 ParsePadded(this string number)
+        public static UInt64 ParsePadded(this string number)
         {
             //Get rid of the leading zeros
             number = number.TrimStart('0');
             if (number == "") return 0;
-            return Int32.Parse(number);
+            return UInt64.Parse(number);
 
         }
